refactor: Create centralized cost database system
- Created costDatabase.ts utility for centralized cost management
- Supports loading costs from localStorage for custom models
- Automatically updates costs for all configured models on app launch
- Removed leftmost separator in bottom menu for cleaner look
- CostTracker now uses centralized database instead of hardcoded costs
- System can now handle any model the user configures, not just hardcoded ones
